+++ src/main/java/org/example/finalbe/domains/rack/dto/EquipmentPlacementRequest.java
@@ -2,11 +2,13 @@ package org.example.finalbe.domains.rack.dto;
 
 import lombok.Builder;
 
+import java.math.BigDecimal;
+
 @Builder
 public record EquipmentPlacementRequest(
         Integer startUnit,
         Integer unitSize,
-        Double powerConsumption,
-        Double weight
+        BigDecimal powerConsumption,
+        BigDecimal weight
 ) {
 }
